Added testing module
