+++ impl/src/planning.rs
@@ -1101,66 +1101,11 @@ mod tests {
     metadata::{CargoMetadataFetcher, Metadata, MetadataFetcher},
     planning::checks,
     settings::testing as settings_testing,
+    testing::*,
   };
 
   use super::*;
   use semver::Version;
-  use std::fs::File;
-  use std::io::Write;
-  use tempfile::TempDir;
-
-  fn basic_toml() -> &'static str {
-    "
-[package]
-name = \"test\"
-version = \"0.0.1\"
-
-[lib]
-path = \"not_a_file.rs\"
-    "
-  }
-
-  fn basic_lock() -> &'static str {
-    "
-[[package]]
-name = \"test\"
-version = \"0.0.1\"
-dependencies = [
-]
-    "
-  }
-
-  fn make_workspace(
-    toml_file: &'static str,
-    lock_file: Option<&'static str>,
-  ) -> (TempDir, CargoWorkspaceFiles) {
-    let dir = TempDir::new().unwrap();
-    let toml_path = {
-      let path = dir.path().join("Cargo.toml");
-      let mut toml = File::create(&path).unwrap();
-      toml.write_all(toml_file.as_bytes()).unwrap();
-      path
-    };
-    let lock_path = match lock_file {
-      Some(lock_file) => {
-        let path = dir.path().join("Cargo.lock");
-        let mut lock = File::create(&path).unwrap();
-        lock.write_all(lock_file.as_bytes()).unwrap();
-        Some(path)
-      },
-      None => None,
-    };
-    let files = CargoWorkspaceFiles {
-      lock_path_opt: lock_path,
-      toml_path,
-    };
-
-    (dir, files)
-  }
-
-  fn make_basic_workspace() -> (TempDir, CargoWorkspaceFiles) {
-    make_workspace(basic_toml(), Some(basic_lock()))
-  }
 
   #[test]
   #[allow(non_snake_case)]
